+++ src/Database/PostgreSQL/PQTypes/Model/ColumnType.hs
@@ -3,7 +3,6 @@ module Database.PostgreSQL.PQTypes.Model.ColumnType (
   , columnTypeToSQL
   ) where
 
-import Control.Applicative ((<$>))
 import Data.Monoid
 import Database.PostgreSQL.PQTypes
 import Prelude