+++ src/Database/PostgreSQL/PQTypes/Model/PrimaryKey.hs
@@ -8,7 +8,6 @@ module Database.PostgreSQL.PQTypes.Model.PrimaryKey (
   , sqlDropPK
   ) where
 
-import Data.Monoid (mconcat)
 import Data.Monoid.Utils
 import Database.PostgreSQL.PQTypes
 import Prelude